feat: Adds document click listener and new options.
This commit introduces the document click listener.

The event, element, screen and viewport data are now
gathered every time a click is made in the page.

Due to this, users can also disable this feature by
setting options.click to false.

Also, there is quite some errors we cannot yet catch,
such as TypeErrors in external scripts (CORS invalid).

For those, we are currently giving the option
for the user to ignore them through
setting `options.ignoreScriptErrors` to false.