feat: Adds plugin support and enhanced Preview
This commit adds plugins support to Perry.

Now it accepts an options.plugins array property.

Plugins are just plain functions that are executed
when the user submits the report.

Plugins are functions that receive the report as the only argument.

Removed the options.credentials prop in order to make
it single purposed.